refactor(rust): move `error_test.rs` into `error.rs` for project consistency
