+++ implementations/rust/ockam/ockam_transport_core/src/lib.rs
@@ -3,5 +3,3 @@
 pub use error::TransportError;
 
 mod error;
-#[cfg(test)]
-mod error_test;
